+++ bin.js
@@ -1,18 +1,16 @@
 #!/usr/bin/env node
 
-var base64 = require('base64-url')
+const base64 = require('base64-url')
 
-var command = process.argv[2]
-var input = process.argv[3]
+const command = process.argv[2]
+const input = process.argv[3]
 
-function commandIsAllowed(c) {
-	return typeof base64[command] === 'function'
-}
+const commandIsAllowed = typeof base64[command] === 'function'
 
-if (!command || !input || !commandIsAllowed(command)) {
-	console.log('Missing required parameters. Run like:')
+if (!command || !input || !commandIsAllowed) {
+	console.log('Missing required parameters, or command is invalid. Run like:')
 	console.log('base64url [encode|decode|escape|unescape] [input]')
 	console.log('E.g., run this: base64url decode Tm9kZS5qcyBpcyBhd2Vzb21lLg')
 } else {
-	console.log(base64[command](input))
+	process.stdout.write(base64[command](input))
 }